Version 1 con colores en interfaces (sin vistas de login)
