--- app/src/main/java/janettha/activity1/Act1/SlidingTapStrip.java
@@ -1,8 +1,0 @@
-package janettha.activity1.Act1;
-
-/**
- * Created by janeth on 2018-02-22.
- */
-
-public class SlidingTapStrip {
-}
