+++ app/src/main/res/layout/dialog_act0.xml
@@ -1,31 +1,38 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout
+    android:id="@+id/lLaCT0"
     xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="wrap_content"
+    android:layout_width="200dp"
     android:layout_height="wrap_content"
     android:orientation="vertical"
     android:layout_gravity="center"
+    android:background="@color/BackgDialogs"
     >
-    <ImageView
-        android:layout_width="100dp"
-        android:layout_height="100dp"
-        android:id="@+id/imgRespuesta"
-        android:background="@drawable/f0"
-        android:layout_gravity="center"/>
     <TextView
         android:id="@+id/nameRespuesta"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="Emocion"
+        android:layout_gravity="center"
+        android:textStyle="bold"
+        android:textSize="@dimen/dialogText"
+        android:textColor="@color/colorTxt"
+        android:layout_margin="@dimen/activity_margin"/>
+    <ImageView
+        android:layout_width="150dp"
+        android:layout_height="150dp"
+        android:id="@+id/imgRespuesta"
+        android:background="@drawable/f0"
+        android:layout_margin="@dimen/activity_margin"
         android:layout_gravity="center"/>
     <Button
         android:id="@+id/btnBack"
-        android:layout_width="150dp"
-        android:layout_height="50dp"
+        android:layout_width="wrap_content"
+        android:layout_height="40dp"
         android:layout_gravity="center"
-        android:text="Ya casi lo logras"
+        android:text="Regresar"
         android:layout_margin="@dimen/activity_margin"
-        android:background="#ffef3b"
+        android:background="#ffd53b"
         />
 
 </LinearLayout>
