+++ app/src/main/res/layout/dialog_act1.xml
@@ -1,11 +1,25 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout
+    android:id="@+id/LLDialogA1"
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
+    android:layout_margin="@dimen/activity_horizontal_margin"
     android:orientation="vertical"
     android:layout_gravity="center"
+    android:background="@color/BackgDialogs"
     >
+    <TextView
+        android:id="@+id/nameRespuesta"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="EMOCION"
+        android:textAlignment="center"
+        android:layout_gravity="center"
+        android:textColor="@color/colorTxt"
+        android:textStyle="bold"
+        />
+
     <TextView
         android:id="@+id/Explicacion"
         android:layout_width="200dp"
@@ -13,8 +27,11 @@
         android:text="¿PREGUNTA?"
         android:textAlignment="center"
         android:textStyle="bold"
+        android:textSize="@dimen/dialogText"
+        android:textColor="@color/colorTxt"
         android:layout_gravity="center_horizontal"
         android:layout_margin="@dimen/activity_margin"
+        android:layout_below="@id/nameRespuesta"
         />
     <ImageView
         android:layout_width="100dp"
@@ -22,20 +39,15 @@
         android:id="@+id/imgRespuesta"
         android:background="@drawable/f0"
         android:layout_gravity="center"/>
-    <TextView
-        android:id="@+id/nameRespuesta"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Emocion"
-        android:layout_gravity="center"/>
     <Button
         android:id="@+id/btnBack"
-        android:layout_width="150dp"
-        android:layout_height="50dp"
+        android:layout_width="wrap_content"
+        android:layout_height="40dp"
         android:layout_gravity="center"
         android:text="Regresar"
+        android:textAlignment="center"
         android:layout_margin="@dimen/activity_margin"
-        android:background="#ffef3b"
+        android:background="#ffd53b"
         />
 
 </LinearLayout>
