+++ app/src/main/res/layout/dialog_act2.xml
@@ -3,29 +3,35 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
+    android:id="@+id/llact2"
     android:orientation="vertical"
     android:layout_gravity="center"
+    android:background="@color/BackgDialogs"
     >
     <TextView
         android:layout_width="200dp"
         android:layout_height="wrap_content"
-        android:text="Toca y manten presionado el botón para grabar una historia."
+        android:text="Toca y manten presionado el botón para grabar una historia que recuerdes acerca de cuando sientes:"
         android:textAlignment="center"
-        android:textStyle="bold"
+        android:textColor="@color/colorTxt"
         android:layout_gravity="center_horizontal"
         android:layout_margin="@dimen/activity_margin"
+        android:textStyle="italic"
     />
-    <ImageView
-        android:layout_width="100dp"
-        android:layout_height="100dp"
-        android:id="@+id/imgEmocion"
-        android:background="@drawable/f0"
-        android:layout_gravity="center"/>
     <TextView
         android:id="@+id/nameEmocion"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="Emocion"
+        android:textColor="@color/colorTxt"
+        android:textStyle="bold"
+        android:layout_margin="@dimen/activity_margin"
+        android:layout_gravity="center"/>
+    <ImageView
+        android:layout_width="100dp"
+        android:layout_height="100dp"
+        android:id="@+id/imgEmocion"
+        android:background="@drawable/f0"
         android:layout_gravity="center"/>
     <LinearLayout
         android:layout_width="wrap_content"
