+++ app/src/main/res/values/colors.xml
@@ -4,9 +4,20 @@
     <color name="colorPrimaryDark">#555959</color>
     <color name="colorAccent">#009e9e</color>
     <color name="color1">#37bdb4</color>
-    <color name="color2">#ea5265</color>
-    <color name="color3">#f4b61a</color>
+    <color name="colorTxt">#9b6431</color>
+    <color name="txtDialogs">#f4b61a</color>
+    <color name="BackgDialogs">#fffdd2</color>
+    <color name="Act0">#5cbfde</color>
+    <color name="Act0Clear">#abd6df</color>
+    <color name="Act1">#fff184</color>
+    <color name="Act1Clear">#b5fff4a0</color>
+    <color name="Act2">#d491fd38</color>
+    <color name="Act2Clear">#d4bbff84</color>
+    <color name="Correcto">#fc2ab633</color>
+    <color name="Incorrecto">#f4594e</color>
 
+
+    <!-- LOGIN -->
     <color name="bg_login">#26ae90</color>
     <color name="bg_register">#2e3237</color>
     <color name="bg_main">#428bca</color>
