+++ app/src/main/res/values/dimens.xml
@@ -2,10 +2,15 @@
     <!-- Default screen margins, per the Android Design guidelines. -->
     <dimen name="activity_margin">8dp</dimen>
     <dimen name="activity_horizontal_margin">16dp</dimen>
-    <dimen name="logo_w_h">80dp</dimen>
     <dimen name="activity_vertical_margin">16dp</dimen>
+    <dimen name="dialogText">15dp</dimen>
+    <dimen name="logo_w_h">80dp</dimen>
     <dimen name="fab_margin">16dp</dimen>
     <dimen name="appbar_padding_top">8dp</dimen>
     <dimen name="app_bar_height">180dp</dimen>
     <dimen name="text_margin">8dp</dimen>
+    <!-- Actividad 1 -->
+    <dimen name="imgLayout_h">380dp</dimen>
+    <dimen name="imgLayout_w">380dp</dimen>
+    <dimen name="imgAct1">80dp</dimen>
 </resources>
